+++ tests/test_metadata_standard_iso19115.py
@@ -1273,13 +1273,3 @@ class InspireMinimalMetadataRecordTestCase(MinimalMetadataRecordTestCase):
             configuration_name='inspire-minimal',
             configuration=config.iso_19115_v1_inspire_v1_3_minimal_record
         )
-
-
-class GeminiCompleteMetadataRecordTestCase(MinimalMetadataRecordTestCase):
-    def setUp(self):
-        super().setUp()
-
-        self._set_metadata_config(
-            configuration_name='gemini-complete',
-            configuration=config.iso_19115_v1_gemini_complete_record
-        )