+++ CHANGELOG.md
@@ -13,7 +13,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* minimal Inspire record configuration (validated against Inspire validator)
 * support for custom access and use constraints
 * support for `gmd:useLimitation` elements
 * CrossRef support for DOI citations
@@ -30,6 +29,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * correcting ISO 19115 service namespace to use ISO hosted schema rather an Inspire draft
 * correcting ISO 19115 namespaces to use HTTPS endpoints
 * correcting test for non-Anchor copyright elements
+
+### Changed
+
+* replaced Gemini typical record configuration with minimal Inspire configuration (validated against Inspire validator)
 
 ## [0.1.1] 2019-07-18